Modified the testing structure
