+++ CMakeLists.txt
@@ -59,7 +59,8 @@ add_compile_definitions(
     )
 
 # This is where your project header files are located.
-include_directories(${CMAKE_CURRENT_LIST_DIR}/include)
+# include_directories(${CMAKE_SOURCE_DIR}/include)
+include_directories(${CMAKE_SOURCE_DIR}/include)
 
 # The main executable is defined inside the src directory.
 add_subdirectory(src)
