+++ src/hello_freertos.c
@@ -6,27 +6,26 @@
 
 #include <stdio.h>
 
-#include "FreeRTOS.h"
+#include "FreeRTOS.h"    
 #include "task.h"
 
-#include "pico/stdlib.h"
+#include "pico/stdlib.h"  
 #include "pico/multicore.h"
 #include "pico/cyw43_arch.h"
+#include "main_functions.h"
 
-int count = 0;
+int count = 0; 
 bool on = false;
 
 #define MAIN_TASK_PRIORITY      ( tskIDLE_PRIORITY + 1UL )
 #define BLINK_TASK_PRIORITY     ( tskIDLE_PRIORITY + 2UL )
-#define MAIN_TASK_STACK_SIZE configMINIMAL_STACK_SIZE
-#define BLINK_TASK_STACK_SIZE configMINIMAL_STACK_SIZE
+
+#define MAIN_TASK_STACK_SIZE    configMINIMAL_STACK_SIZE
+#define BLINK_TASK_STACK_SIZE   configMINIMAL_STACK_SIZE
 
 void blink_task(__unused void *params) {
-    hard_assert(cyw43_arch_init() == PICO_OK);
     while (true) {
-        cyw43_arch_gpio_put(CYW43_WL_GPIO_LED_PIN, on);
-        if (count++ % 11) on = !on;
-        vTaskDelay(500);
+        led_blink(on, count);
     }
 }
 
@@ -35,20 +34,21 @@ void main_task(__unused void *params) {
                 BLINK_TASK_STACK_SIZE, NULL, BLINK_TASK_PRIORITY, NULL);
     char c;
     while(c = getchar()) {
-        if (c <= 'z' && c >= 'a') putchar(c - 32);
-        else if (c >= 'A' && c <= 'Z') putchar(c + 32);
-        else putchar(c);
+        char_inversion(c);
     }
 }
 
-int main( void )
-{
+int main(void) {
     stdio_init_all();
     const char *rtos_name;
     rtos_name = "FreeRTOS";
     TaskHandle_t task;
+    
+    // xTaskCreate, taking the function name, task size and task priority 
     xTaskCreate(main_task, "MainThread",
                 MAIN_TASK_STACK_SIZE, NULL, MAIN_TASK_PRIORITY, &task);
+
     vTaskStartScheduler();
+    
     return 0;
 }
