+++ test/test.c
@@ -4,6 +4,7 @@
 #include <unity.h>
 #include "unity_config.h"
 #include "pico/cyw43_arch.h"
+#include "main_functions.h"
 
 void setUp(void) {}
 
@@ -23,46 +24,36 @@ void test_multiplication(void)
     TEST_ASSERT_TRUE_MESSAGE(z == 5, "Multiplication of two integers returned incorrect value.");
 }
 
-char read_char(char c){
-    if (c <= 'z' && c >= 'a') putchar(c - 32);
-    else if (c >= 'A' && c <= 'Z') putchar(c + 32);
-    else putchar(c);
-    return c;
-}
-
 void test_led_pin(){
     int count = 0;
-    bool on = false;    
+    bool on = false;
+    bool prev_state = false;   
     int gpio_state = false;
     int check_count = 0;
     hard_assert(cyw43_arch_init() == PICO_OK);
-    // Testing 10 times as 110 is multiple of 11 and checking how many times it enters the loop to verify.
     for(int loop_count = 0; loop_count <111; loop_count ++)
     {
-        cyw43_arch_gpio_put(CYW43_WL_GPIO_LED_PIN, on);
+        LedBlinkResult result = led_blink(on, count);
         gpio_state = cyw43_arch_gpio_get(CYW43_WL_GPIO_LED_PIN);
-        // Reading GPIO state and checking if the valuewe set is equal to value read back from board.
-        TEST_ASSERT_EQUAL_MESSAGE(gpio_state, on, "TEST FAILED");
-        if(count++ % 11){
-            check_count++;
-            on = !on;
+        TEST_ASSERT_EQUAL_MESSAGE(gpio_state, result.on, "TEST FAILED");
+        if ((result.count - 1) % 11){
+            TEST_ASSERT_NOT_EQUAL_MESSAGE(gpio_state, prev_state, "TEST FAILED");
         }
+        prev_state = gpio_state;
         sleep_ms(500);
     }
-    // We are verifying that `check_count` equals 100 because, according to the loop logic, it should execute exactly 100 times—neither more nor less.
-    TEST_ASSERT_EQUAL_MESSAGE(check_count, 100, "TEST FAILED");
 }
 
 void test_read_char(){
     char result;
     //For lower case letters
     for (char c = 'a'; c<='z'; c++){
-    result = read_char(c);
+    result = char_inversion(c);
     TEST_ASSERT_TRUE_MESSAGE(result = c - 32 , "TEST FAILED");
     }
     //For Upper case letters
     for (char c = 'A'; c<='Z'; c++){
-    result = read_char(c);
+    result = char_inversion(c);
     TEST_ASSERT_TRUE_MESSAGE(result = c + 32 , "TEST FAILED");
     }
 }
@@ -71,10 +62,8 @@ int main (void)
 {
     stdio_init_all();
     sleep_ms(5000); // Give time for TTY to attach.
-    printf("Start tests\n");
+    printf("BEGIN TESTING\n");
     UNITY_BEGIN();
-    // RUN_TEST(test_variable_assignment);
-    // RUN_TEST(test_multiplication);
     RUN_TEST(test_read_char);
     RUN_TEST(test_led_pin);
     sleep_ms(5000);
